pyproject.toml
[tool.poetry]
name = "models"
version = "0.1.0"
description = ""
authors = ["petereon <pvyboch1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
solidpython = "^1.1.3"
watchdog = "^2.1.9"
poethepoet = "^0.16.2"
pymeshfix = "^0.16.1"

[tool.poe]
verbosity=-1

[tool.poe.tasks]
"render" = { shell = "poetry run python ${FILE_PATH}", args = [{ name = "FILE_PATH", positional = true}] }
"render:watch" = { shell = "poetry run python ${FILE_PATH} ; poetry run watchmedo shell-command --patterns='*.py;*.feature;*.toml' --recursive --drop --command='poetry run python ${FILE_PATH}'", args = [{ name = "FILE_PATH", positional = true}] }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
